+++ src/port/ui/ResolutionEditor.cpp
@@ -115,7 +115,7 @@ void ResolutionCustomWidget(WidgetInfo& info) {
         verticalPixelCount = pixelCountPresets[item_pixelCount];
 
         if (showHorizontalResField) {
-            horizontalPixelCount = (verticalPixelCount / aspectRatioY) * aspectRatioX;
+            horizontalPixelCount = static_cast<int32_t>((verticalPixelCount / aspectRatioY) * aspectRatioX);
         }
 
         CVarSetInteger(CVAR_PREFIX_ADVANCED_RESOLUTION ".VerticalPixelCount", verticalPixelCount);
@@ -134,8 +134,8 @@ void ResolutionCustomWidget(WidgetInfo& info) {
                 if (horizontalPixelCount < SCREEN_WIDTH) {
                     horizontalPixelCount = SCREEN_WIDTH;
                 }
-                aspectRatioX = horizontalPixelCount;
-                aspectRatioY = verticalPixelCount;
+                aspectRatioX = static_cast<float>(horizontalPixelCount);
+                aspectRatioY = static_cast<float>(verticalPixelCount);
                 update[UPDATE_aspectRatioX] = true;
                 update[UPDATE_aspectRatioY] = true;
             }
@@ -151,7 +151,7 @@ void ResolutionCustomWidget(WidgetInfo& info) {
                 aspectRatioY = aspectRatioPresetsY[2];
                 update[UPDATE_aspectRatioX] = true;
                 update[UPDATE_aspectRatioY] = true;
-                horizontalPixelCount = (verticalPixelCount / aspectRatioY) * aspectRatioX;
+                horizontalPixelCount = static_cast<int32_t>((verticalPixelCount / aspectRatioY) * aspectRatioX);
             }
         }
     }
@@ -165,8 +165,8 @@ void ResolutionCustomWidget(WidgetInfo& info) {
         // Ignore vertical resolutions that are below the lower clamp constant.
         if (showHorizontalResField && !(verticalPixelCount < minVerticalPixelCount)) {
             item_aspectRatio = default_aspectRatio;
-            aspectRatioX = horizontalPixelCount;
-            aspectRatioY = verticalPixelCount;
+            aspectRatioX = static_cast<float>(horizontalPixelCount);
+            aspectRatioY = static_cast<float>(verticalPixelCount);
             update[UPDATE_aspectRatioX] = true;
             update[UPDATE_aspectRatioY] = true;
         }
@@ -266,12 +266,12 @@ void ResolutionCustomWidget(WidgetInfo& info) {
             if (!showHorizontalResField && (aspectRatioX > 0.0f)) { // when turning this setting off
                 // Refresh relevant values
                 aspectRatioX = aspectRatioY * horizontalPixelCount / verticalPixelCount;
-                horizontalPixelCount = (verticalPixelCount / aspectRatioY) * aspectRatioX;
+                horizontalPixelCount = static_cast<int32_t>((verticalPixelCount / aspectRatioY) * aspectRatioX);
             } else { // when turning this setting on
                 item_aspectRatio = default_aspectRatio;
                 if (aspectRatioX > 0.0f) {
                     // Refresh relevant values in the opposite order
-                    horizontalPixelCount = (verticalPixelCount / aspectRatioY) * aspectRatioX;
+                    horizontalPixelCount = static_cast<int32_t>((verticalPixelCount / aspectRatioY) * aspectRatioX);
                     aspectRatioX = aspectRatioY * horizontalPixelCount / verticalPixelCount;
                 }
             }
@@ -460,7 +460,7 @@ void RegisterResolutionWidgets() {
                 aspectRatioY = aspectRatioPresetsY[item_aspectRatio];
 
                 if (showHorizontalResField) {
-                    horizontalPixelCount = (verticalPixelCount / aspectRatioY) * aspectRatioX;
+                    horizontalPixelCount = static_cast<int32_t>((verticalPixelCount / aspectRatioY) * aspectRatioX);
                 }
 
                 CVarSetFloat(CVAR_PREFIX_ADVANCED_RESOLUTION ".AspectRatioX", aspectRatioX);
@@ -583,7 +583,7 @@ void UpdateResolutionVars() {
     verticalPixelCount =
         CVarGetInteger(CVAR_PREFIX_ADVANCED_RESOLUTION ".VerticalPixelCount", pixelCountPresets[item_pixelCount]);
     // Additional settings
-    horizontalPixelCount = (verticalPixelCount / aspectRatioY) * aspectRatioX;
+    horizontalPixelCount = static_cast<int32_t>((verticalPixelCount / aspectRatioY) * aspectRatioX);
     // Disabling flags
     disabled_everything = !CVarGetInteger(CVAR_PREFIX_ADVANCED_RESOLUTION ".Enabled", 0);
     disabled_pixelCount = !CVarGetInteger(CVAR_PREFIX_ADVANCED_RESOLUTION ".VerticalResolutionToggle", 0);
